backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9
pydantic[email]==2.6.1
python-dotenv==1.0.1
stripe>=9.0.0
redis==5.0.1
celery==5.3.6
authlib==1.3.0
httpx==0.27.0
sentry-sdk[fastapi]==1.40.6
slowapi==0.1.9
psycopg2-binary==2.9.11
docker==7.0.0
requests==2.31.0
uvloop==0.19.0
alembic==1.13.1
email-validator==2.1.1
structlog==24.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
